+++ src/layer_utils/vector_layer.py
@@ -23,7 +23,7 @@ def checkVectorLayer(layer):
         return True, ""
 
 
-def checkVectorLayerVelocity(layer):
+def getVectorVelocityFieldName(layer):
     """ check layer is a valid vector with velocity """
 
     velocity_field_name_options = ['velocity', 'VEL', 'mean_velocity']
@@ -97,3 +97,9 @@ def extractDateValueAttributes(attributes: dict) -> list:
             date_value_list.append((date_obj, float(value)))
 
     return np.array(date_value_list, dtype=object)
+
+
+def getVectorFields(layer):
+    """ get field names from vector layer"""
+    fields = [field.name() for field in layer.fields()]
+    return fields